requirements.txt
Django==3.2
django-extensions==3.1.5
djangorestframework==3.13.1
django-cors-headers
psycopg2-binary==2.8
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
Pillow==9.0.0
python-dateutil==2.8.2
django-fs-trumbowyg @ git+https://github.com/jon-alamo/django-fs-trumbowyg.git@7dc233d53713ca6724fdf29fb1c7d0c2970f1adc
PyYAML==6.0
django-location-field==2.1.0
envs==1.3
stripe==2.65.0
django-inlinecss == 0.3.0
gunicorn==20.1.0
django-anymail[mailgun]==8.5
django-modeltranslation==0.17.5
django-background-tasks==1.2.5
requests==2.27.1
django-naomi==0.8
django-localflavor==3.1
qrcode[pil]==7.3.1
pdfkit==1.0.0
weasyprint==52.5
django-non-dark-admin==2.0.2
Faker
django-redis
